backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation and settings
pydantic==2.7.4
pydantic-settings==2.3.3
email-validator==2.1.0

# HTTP client
httpx==0.27.0

# Blockchain integration
web3==6.11.3
tronpy==0.4.0

# Bitcoin integration
bitcoinlib==0.12.0
electrum==4.4.6

# Caching and sessions
redis==5.0.1
aioredis==2.0.1

# Cryptography
cryptography==41.0.8
fernet==1.0.1

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1
bandit==1.7.5

# Monitoring and logging
structlog==23.2.0
